Refactor export win functional tests
