+++ src/client/modules/ExportWins/Form/validators.js
@@ -1,4 +1,4 @@
-import { isDateWithinLastTwelveMonths } from './utils'
+import { isWithinLastTwelveMonths } from './utils'
 
 export const POSITIVE_INT_REGEX = /^[0-9]{1,19}$/
 export const isPositiveInteger = (value) => POSITIVE_INT_REGEX.test(value)
@@ -9,6 +9,6 @@ export const validateTeamMembers = (team_members) =>
     : null
 
 export const validateWinDate = ({ month, year }) =>
-  isDateWithinLastTwelveMonths(new Date(year, month - 1))
+  isWithinLastTwelveMonths(new Date(year, month - 1))
     ? null
     : 'Date must be in the last 12 months'
