+++ test/functional/cypress/fakers/export-wins.js
@@ -1,5 +1,7 @@
 import { faker } from '@faker-js/faker'
 
+import { contactFaker } from './contacts'
+
 /**
  * Generate fake data for a single export win.
  */
@@ -20,6 +22,7 @@ export const exportWinsFaker = () => ({
   customer_name: faker.person.fullName(),
   customer_job_title: faker.person.jobTitle(),
   customer_email_address: faker.internet.email(),
+  company_contacts: [contactFaker()],
   total_expected_export_value: faker.number.int({
     min: 10_000,
     max: 10_000_000,
